pyproject.toml
[project]
name = "sber-annotations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.6.0",
    "loguru>=0.7.3",
    "mlflow>=2.22.0",
    "opencv-python>=4.11.0.86",
    "qwen-vl-utils>=0.0.11",
    "torchvision>=0.22.0",
    "transformers[torch]>=4.51.3",
    "ultralytics>=8.3.132",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pre-commit>=4.2.0",
    "ruff>=0.11.10",
]
